Reduce float and sampler precision where not critical

+++ src/shaders/solve_pressure.frag
@@ -1,6 +1,6 @@
 #version 300 es
-precision highp float;
-precision highp sampler2D;
+precision mediump float;
+precision mediump sampler2D;
 
 uniform float rBeta;
 uniform float alpha;
